+++ app/src/main/java/com/example/unbegrenzt/fharmaapp/Fragments/Map.java
@@ -1,8 +1,8 @@
 /*
- * Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-03-17 09:54 PM
+ * Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-10-17 11:31 PM
  * Copyright (c) 2017. All rights reserved.
  *
- * Last modified 06-03-17 05:05 PM
+ * Last modified 06-10-17 11:31 PM
  */
 
 package com.example.unbegrenzt.fharmaapp.Fragments;
@@ -17,16 +17,15 @@ import android.view.ViewGroup;
 
 import com.example.unbegrenzt.fharmaapp.R;
 
-
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link farmacos.OnFragmentInteractionListener} interface
+ * {@link Map.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link farmacos#newInstance} factory method to
+ * Use the {@link Map#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class farmacos extends Fragment {
+public class Map extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -38,7 +37,7 @@ public class farmacos extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public farmacos() {
+    public Map() {
         // Required empty public constructor
     }
 
@@ -48,11 +47,11 @@ public class farmacos extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment farmacos.
+     * @return A new instance of fragment Map.
      */
     // TODO: Rename and change types and number of parameters
-    public static farmacos newInstance(String param1, String param2) {
-        farmacos fragment = new farmacos();
+    public static Map newInstance(String param1, String param2) {
+        Map fragment = new Map();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -73,18 +72,7 @@ public class farmacos extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        /*View rootView = inflater.inflate(R.layout.farmacos, container, false);
-
-        RecyclerView rv = (RecyclerView) rootView.findViewById(R.id.rv_recycler_view);
-        rv.setHasFixedSize(true);
-        MyAdapter adapter = new MyAdapter(new String[]{"Example One", "Example Two", "Example Three", "Example Four", "Example Five" , "Example Six" , "Example Seven"});
-        rv.setAdapter(adapter);
-
-        LinearLayoutManager llm = new LinearLayoutManager(getActivity());
-        rv.setLayoutManager(llm);
-
-        return rootView;*/
-        return inflater.inflate(R.layout.farmacos, container, false);
+        return inflater.inflate(R.layout.map, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
