--- app/src/main/res/drawable-v21/ripple_effect_second.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-04-17 07:50 PM
-  ~ Copyright (c) 2017. All rights reserved.
-  ~
-  ~ Last modified 06-04-17 07:50 PM
-  -->
-
-<ripple xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:color="@color/atenuante_ripple1"
-    tools:targetApi="lollipop">
-    <item android:id="@android:id/mask">
-        <shape android:shape="rectangle">
-            <solid android:color="@color/atenuante_ripple2" />
-        </shape>
-    </item>
-    <item
-        android:id="@android:id/background"
-        android:drawable="@color/atenuante" />
-</ripple>
