--- app/src/main/res/drawable/button_pressed.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
-    <solid android:color="@color/buttonColorPressed"/>
-
-    <stroke android:width="3dp"
-        android:color="@color/buttonColorPressed"/>
-
-    <padding android:left="1dp"
-        android:top="1dp"
-        android:right="1dp"
-        android:bottom="1dp"/>
-
-    <corners android:radius="@dimen/borde_boton"/>
-</shape>
