--- app/src/main/res/drawable/circle.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:innerRadius="0dp"
-    android:shape="ring"
-    android:thicknessRatio="2.0"
-    android:useLevel="false" >
-    <solid android:color="@android:color/transparent" />
-
-    <stroke
-        android:width="8dp"
-        android:color="@android:color/white" />
-</shape>
