--- app/src/main/res/drawable/login_back.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-04-17 04:39 PM
-  ~ Copyright (c) 2017. All rights reserved.
-  ~
-  ~ Last modified 06-04-17 04:39 PM
-  -->
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <gradient
-        android:angle="135"
-        android:centerColor="#81C784"
-        android:endColor="@color/acentuado_oscuro"
-        android:startColor="@color/color_acentuado"
-        android:type="linear"/>
-
-</shape>
