--- app/src/main/res/drawable/ripple_effect_second.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-04-17 07:58 PM
-  ~ Copyright (c) 2017. All rights reserved.
-  ~
-  ~ Last modified 06-04-17 07:58 PM
-  -->
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_selected="true" android:drawable="@color/atenuante"/>
-    <item android:state_pressed="true" android:drawable="@color/atenuante_ripple1"/>
-    <item android:drawable="@drawable/button_normal"/>
-</selector>
