+++ app/src/main/res/layout/farmacias.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/color_base"
     tools:context="com.example.unbegrenzt.fharmaapp.Fragments.Farmacias">
 
     <android.support.v4.widget.SwipeRefreshLayout
