+++ app/src/main/res/layout/menu.xml
@@ -6,7 +6,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:openDrawer="start">
+    tools:openDrawer="start"
+    android:background="@color/color_de_texto">
 
     <include
         layout="@layout/app_bar_menu"
