--- app/src/main/res/layout/signup_design.xml
@@ -1,171 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-04-17 08:23 PM
-  ~ Copyright (c) 2017. All rights reserved.
-  ~
-  ~ Last modified 06-04-17 08:23 PM
-  -->
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <ImageView
-        android:id="@+id/imagen_upload"
-        android:layout_marginTop="20dp"
-        android:layout_centerHorizontal="true"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:background="@drawable/selecteditem_dot"
-        android:src="@drawable/upload_photo"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="@dimen/margen_izq"
-        android:layout_below="@+id/imagen_upload"
-        android:layout_marginTop="30dp"
-        android:text="@string/nete_a_trav_z_de"
-        android:textColor="@color/color_de_texto"
-        android:textSize="@dimen/texto_subhead1_3A"
-        android:id="@+id/txt2"/>
-
-    <LinearLayout
-        android:layout_below="@+id/txt2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="5dp"
-        android:id="@+id/lln"
-        android:layout_marginLeft="@dimen/margen_izq"
-        android:layout_marginRight="@dimen/margen_izq"
-        android:weightSum="2.2">
-
-        <android.support.v7.widget.AppCompatButton
-
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@drawable/ripple_effect"
-            android:id="@+id/facebooksignn"
-            android:textSize="@dimen/texto_subhead2_3B"
-            android:text="@string/facebook"
-            android:layout_weight="1"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="0.2"/>
-
-        <android.support.v7.widget.AppCompatButton
-
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@color/atenuante"
-            android:id="@+id/googlesign"
-            android:textSize="@dimen/texto_subhead2_3B"
-            android:text="@string/google"
-            android:layout_weight="1"/>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_below="@+id/lln"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="15dp"
-        android:id="@+id/lln2"
-        android:layout_marginLeft="@dimen/margen_izq"
-        android:layout_marginRight="@dimen/margen_izq"
-        android:weightSum="2.2">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/or"
-            android:id="@+id/bel"
-            android:textAlignment="center"
-            android:layout_weight="0.2"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"/>
-
-
-    </LinearLayout>
-
-
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/enter_name"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/lln2"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:layout_marginLeft="@dimen/margen_izq"
-        android:layout_marginRight="@dimen/margen_izq">
-
-        <EditText android:id="@+id/input_name"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:inputType="textEmailAddress"
-            android:hint="@string/nombre_de_usuario" />
-
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/enter_emailtodo"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/enter_name"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:layout_marginLeft="@dimen/margen_izq"
-        android:layout_marginRight="@dimen/margen_izq">
-
-        <EditText android:id="@+id/enter_email"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:inputType="textEmailAddress"
-            android:hint="@string/email" />
-
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.design.widget.TextInputLayout
-        android:id="@+id/enter_contra"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:layout_below="@+id/enter_emailtodo"
-        android:layout_marginLeft="@dimen/margen_izq"
-        android:layout_marginRight="@dimen/margen_izq">
-
-        <EditText android:id="@+id/enter_password"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:inputType="textPassword"
-            android:hint="@string/password"/>
-
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.v7.widget.AppCompatButton
-        android:layout_below="@+id/enter_contra"
-        android:id="@+id/link_signup"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:background="@drawable/ripple_effect_second"
-        android:textColor="@color/color_de_texto"
-        android:text="@string/crear_cuenta"
-        android:textSize="@dimen/texto_subhead2_3B"
-        android:layout_marginLeft="@dimen/margen_izq"
-        android:layout_marginRight="@dimen/margen_izq"/>
-
-</RelativeLayout>
