+++ app/src/main/res/layout/vista.xml
@@ -39,25 +39,17 @@
 
     <Button
         android:id="@+id/BotonInicio"
-        android:layout_width="300dp"
-        android:drawableLeft="@drawable/fb_logo"
+        android:layout_width="100dp"
         android:layout_height="wrap_content"
         android:layout_above="@+id/dot2"
         android:layout_centerInParent="true"
-        android:text="@string/iniciar_sesion_con_facebook"
+        android:text="@string/continuar"
         android:textAllCaps="false"
         android:background="@drawable/ripple_effect_face"
         android:layout_marginBottom="@dimen/dot_marginbotom"
         android:textSize="@dimen/text_face"
         android:onClick="start"/>
 
-    <com.facebook.login.widget.LoginButton
-        android:id="@+id/login_button"
-        android:layout_width="wrap_content"
-        android:layout_height="90dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="30dp"
-        android:visibility="gone"/>
 
     <ImageView
         android:id="@+id/dot1"
