+++ app/src/main/res/values-v23/styles.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~ Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-03-17 12:29 PM
+  ~ Created  by Unbegrenzt for Jorge Luis Morales Centeno on 06-11-17 12:49 PM
   ~ Copyright (c) 2017. All rights reserved.
   ~
-  ~ Last modified 04-10-17 02:31 PM
+  ~ Last modified 06-11-17 12:48 PM
   -->
 
 <resources>
