+++ app/src/main/res/values/dimen.xml
@@ -1,13 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="texto_header_1B">45sp</dimen>
-    <dimen name="texto_header_1A">34sp</dimen>
     <dimen name="texto_headline_2">24sp</dimen>
     <dimen name="text_face">18sp</dimen>
-    <dimen name="texto_subhead2_3B">16sp</dimen>
     <dimen name="texto_subhead1_3A">15sp</dimen>
-    <dimen name="texto_body2_4B">14sp</dimen>
-    <dimen name="texto_body1_4A">13sp</dimen>
 
     <!--1B con 3A, 2 con 4A, 1A con 3A, 2A con 4A -->
 </resources>
