+++ app/src/main/res/values/styles.xml
@@ -8,19 +8,23 @@
 <resources>
 
     <!-- Base application theme. -->
-    <style name="AppTheme" parent="Theme.AppCompat.NoActionBar">
+    <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
         <!-- Customize your theme here. -->
 
-        <item name="colorPrimary">@color/color_base</item>
-        <item name="colorPrimaryDark">@color/complemento_Acentuado</item>
-        <item name="colorAccent">@color/color_acentuado</item>
-        <item name="android:colorForeground">@color/atenuante</item>
-        <item name="android:textColorPrimary">@color/color_de_texto</item>
-        <item name="android:textColorSecondaryInverse">@android:color/black</item>
+        <item name="colorPrimary">@color/primary1</item>
+        <item name="colorPrimaryDark">@color/primary_dark1</item>
+        <item name="colorAccent">@color/accent1</item>
+        <item name="android:colorForeground">@color/primary_dark1</item>
+        <item name="android:textColorPrimary">@color/primary_text1</item>
+        <item name="android:textColorSecondaryInverse">@color/divider1</item>
+        <item name="android:textColorSecondary">@color/secondary_text1</item>
+        <item name="android:colorBackground">@color/icons1</item>
+        <item name="android:textColorPrimaryInverse">@color/primary_text1</item>
+        <item name="android:windowBackground">?android:attr/colorBackground</item>
 
     </style>
 
-    <style name="AppTheme.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+    <style name="AppTheme.AppBarOverlay" parent="Base.ThemeOverlay.AppCompat.ActionBar" />
 
     <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 
@@ -29,7 +33,7 @@
         <item name="windowNoTitle">true</item>
     </style>
 
-    <style name="SplashTheme" parent="Theme.AppCompat.NoActionBar">
+    <style name="SplashTheme" parent="AppTheme">
         <item name="android:windowBackground">@drawable/background_splash</item>
     </style>
